etl trait based rework
