--- src/connection/etl/import/trim.rs
@@ -1,17 +1,0 @@
-/// Trim options for IMPORT
-#[derive(Debug, Clone, Copy)]
-pub enum Trim {
-    Left,
-    Right,
-    Both,
-}
-
-impl AsRef<str> for Trim {
-    fn as_ref(&self) -> &str {
-        match self {
-            Self::Left => "LTRIM",
-            Self::Right => "RTRIM",
-            Self::Both => "TRIM",
-        }
-    }
-}
