--- src/connection/etl/row_separator.rs
@@ -1,16 +1,0 @@
-#[derive(Debug, Clone, Copy)]
-pub enum RowSeparator {
-    LF,
-    CR,
-    CRLF,
-}
-
-impl AsRef<str> for RowSeparator {
-    fn as_ref(&self) -> &str {
-        match self {
-            RowSeparator::LF => "LF",
-            RowSeparator::CR => "CR",
-            RowSeparator::CRLF => "CRLF",
-        }
-    }
-}
