--- src/connection/etl/traits/mod.rs
@@ -1,5 +1,0 @@
-mod job;
-mod with_socket_maker;
-
-pub use job::EtlJob;
-pub use with_socket_maker::WithSocketMaker;
