+++ src/responses/hosts.rs
@@ -6,11 +6,5 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Hosts {
-    nodes: Vec<IpAddr>,
-}
-
-impl From<Hosts> for Vec<IpAddr> {
-    fn from(value: Hosts) -> Self {
-        value.nodes
-    }
+    pub nodes: Vec<IpAddr>,
 }
